saga_api and affected tools: removed redundancies to CSG_Table::Get_Count() => CSG_Table_Get_Record_Count(), CSG_PointCloud::Get_Point_Count(), renamed CSG_Tool_Grid::Set_Progress(int Row) => Set_Progress_Rows(), additional CSG_Table::Set_Value() functions for integer types (sLong, size_t, int), ...

+++ saga-gis/src/tools/grid/grid_calculus/Grid_Random_Terrain.cpp
@@ -125,8 +125,8 @@ bool CGrid_Random_Terrain::On_Execute(void)
 //---------------------------------------------------------
 void CGrid_Random_Terrain::Add_Bump(void)
 {
-	int	x	= CSG_Random::Get_Uniform(-m_Radius, m_Radius + m_pGrid->Get_NX());
-	int	y	= CSG_Random::Get_Uniform(-m_Radius, m_Radius + m_pGrid->Get_NY());
+	int	x	= (int)CSG_Random::Get_Uniform(-m_Radius, m_Radius + m_pGrid->Get_NX());
+	int	y	= (int)CSG_Random::Get_Uniform(-m_Radius, m_Radius + m_pGrid->Get_NY());
 
 	for(int i=0; i<m_Kernel.Get_Count(); i++)
 	{
